svn merge ^/trunk@9240 ^/branches/sync-server-trunk-merge .

This is the "2-URL merge".  Since the branch already contains all the
changes from the trunk, this only merges the differences between current
trunk and the branch into the trunk.  This makes the trunk look just like
the branch.


git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk/nimbus@9241 640d5228-2204-0410-8d71-8f46fa6850b4

+++ src/filter/setup.c
@@ -99,12 +99,12 @@ void LoadSetup_OK(Widget w, XtPointer client, XmFileSelectionBoxCallbackStruct *
 
   if (!ProductionSetup)
   {
-    int value = atoi(strchr(buffer, '=')+1);
+    long value = atoi(strchr(buffer, '=')+1);
     switch (value)
     {
       case 25:
       case 50:
-        SetHighRate(NULL, (void *)value, NULL);
+        SetHighRate(NULL, (void*)value, NULL);
         break;
       case Config::SampleRate:
         SetSampleRate(NULL, NULL, NULL);
